refactor(webapi): replace CodeRepositoryConnectedSchema with CodeRepositoryConnectedRecord for improved event handling and structure
